+++ chromance-firmware/src/animations/centerPulseAnimation.cpp
@@ -1,5 +1,5 @@
 #include "centerPulseAnimation.h"
-#include "chromance/map.h"
+#include "ripples/map.h"
 
 using namespace Chromance;
