+++ chromance-firmware/src/animations/centerPulseAnimation.h
@@ -1,7 +1,7 @@
 #ifndef CENTERPULSE_ANIMATION_H_
 #define CENTERPULSE_ANIMATION_H_
 
-#include "chromance/rippleAnimation.h"
+#include "ripples/rippleAnimation.h"
 
 namespace Chromance
 {
