+++ chromance-firmware/src/animations/cubePulseAnimation.cpp
@@ -1,5 +1,5 @@
 #include "cubePulseAnimation.h"
-#include "chromance/map.h"
+#include "ripples/map.h"
 
 using namespace Chromance;
