+++ chromance-firmware/src/animations/cubePulseAnimation.h
@@ -1,7 +1,7 @@
 #ifndef CUBEPULSE_ANIMATION_H_
 #define CUBEPULSE_ANIMATION_H_
 
-#include "chromance/rippleAnimation.h"
+#include "ripples/rippleAnimation.h"
 
 namespace Chromance
 {
