+++ chromance-firmware/src/animations/ripples/map.h
