+++ chromance-firmware/src/animations/ripples/ripple.cpp
