+++ chromance-firmware/src/animations/ripples/ripple.h
