+++ chromance-firmware/src/animations/ripples/rippleAnimation.cpp
