+++ chromance-firmware/src/animations/ripples/rippleAnimation.h
