+++ chromance-firmware/src/animations/ripples/ripplePool.cpp
