+++ chromance-firmware/src/animations/ripples/ripplePool.h
