+++ chromance-firmware/src/animations/starBurstPulseAnimation.cpp
@@ -1,5 +1,5 @@
 #include "starBurstPulseAnimation.h"
-#include "chromance/map.h"
+#include "ripples/map.h"
 
 using namespace Chromance;
