+++ chromance-firmware/src/animations/starBurstPulseAnimation.h
@@ -1,7 +1,7 @@
 #ifndef STARBURSTPULSE_ANIMATION_H_
 #define STARBURSTPULSE_ANIMATION_H_
 
-#include "chromance/rippleAnimation.h"
+#include "ripples/rippleAnimation.h"
 
 namespace Chromance
 {
